dialog pour suppression d'un quiz

+++ src/app/themes/edit-theme/edit-theme.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { ConfigurationService } from 'src/services/configuration.service';
-import { QUIZZES } from 'src/mocks/quiz.mock';
 import { Quiz } from 'src/models/quiz.model';
 import { QuizService } from 'src/services/quiz.service';
+import { MatDialog, MatDialogConfig, MatDialogRef } from "@angular/material/dialog";
+import { DeleteQuizDialog } from 'src/app/dialogs/delete-quiz/delete-quiz-dialog.component';
 
 @Component({
     selector: 'app-edit-theme',
@@ -13,7 +14,7 @@ export class EditThemeComponent implements OnInit {
 
     quizzes: Quiz[];
 
-    constructor(public configService: ConfigurationService, private quizService: QuizService) {
+    constructor(public configService: ConfigurationService, private quizService: QuizService, private dialog: MatDialog) {
         this.quizService.quizzes$.subscribe((quizzes: Quiz[]) => {
             this.quizzes = quizzes;
         });
@@ -23,6 +24,21 @@ export class EditThemeComponent implements OnInit {
     }
 
     deleteQuiz(quiz: Quiz) {
-        this.quizService.deleteQuiz(quiz);
+        const dialogRef = this.openDialog();
+        dialogRef.afterClosed().subscribe(
+            result => {
+                if (result) this.quizService.deleteQuiz(quiz);
+            }
+        )
+    }
+
+    openDialog(): MatDialogRef<DeleteQuizDialog, any> {
+        const dialogConfig = new MatDialogConfig();
+
+        dialogConfig.disableClose = true;
+
+        dialogConfig.autoFocus = true;
+
+        return this.dialog.open(DeleteQuizDialog, dialogConfig);
     }
 }
